ftofstrip, ctofstrip, cndstrip, pcustrip fixed

+++ src/trigger/cnd.s/cndstrips.cc
@@ -23,8 +23,8 @@ using namespace std;
   chan = ch_m; \
   str    = adcstrip[isl][chan] - 1; \
   lr     = adclr[isl][chan] - 1; \
-  timexxx = ((energy >= strip_threshold) ? ((uint8_t)fadcs.fadc[isl].t##ch_m) : 0); /* error in '?' without (uint16_t) ...*/ \
-  energyxxx = ((energy >= strip_threshold) ? ((uint8_t)fadcs.fadc[isl].e##ch_m) : 0); /* error in '?' without (uint16_t) ...*/ \
+  timexxx = ((energy >= strip_threshold) ? ((uint16_t)fadcs.fadc[isl].t##ch_m) : 0); /* error in '?' without (uint16_t) ...*/ \
+  energyxxx = ((energy >= strip_threshold) ? ((uint16_t)fadcs.fadc[isl].e##ch_m) : 0); /* error in '?' without (uint16_t) ...*/ \
   timetmp[lr][str] = timexxx; \
   energytmp[lr][str] = energyxxx
